+++ src/dcos_e2e/_vendor/dcos_launch/platforms/aws.py
@@ -32,9 +32,9 @@ log = logging.getLogger(__name__)
 
 def template_by_instance_type(instance_type):
     if instance_type.split('.')[0] in ('c4', 't2', 'm4'):
-        template = pkg_resources.resource_string('dcos_launch', 'templates/vpc-ebs-only-cluster-template.json')
+        template = pkg_resources.resource_string('dcos_e2e._vendor.dcos_launch', 'templates/vpc-ebs-only-cluster-template.json')
     else:
-        template = pkg_resources.resource_string('dcos_launch', 'templates/vpc-cluster-template.json')
+        template = pkg_resources.resource_string('dcos_e2e._vendor.dcos_launch', 'templates/vpc-cluster-template.json')
     return template.decode('utf-8')