+++ src/dcos_e2e/_vendor/dcos_launch/util.py
@@ -45,7 +45,7 @@ def stub(output):
 
 def get_temp_config_path(tmpdir, name, update: dict = None):
     config = yaml.load(
-        pkg_resources.resource_string('dcos_launch', 'sample_configs/{}'.format(name)).decode('utf-8'))
+        pkg_resources.resource_string('dcos_e2e._vendor.dcos_launch', 'sample_configs/{}'.format(name)).decode('utf-8'))
     if update is not None:
         config.update(update)
     new_config_path = tmpdir.join('my_config.yaml')